게시판 목록 이쁘게 꾸밈

+++ g_board/src/test/java/com/mycompany/myapp/conntest/mappertest.java
@@ -23,7 +23,7 @@ public class mappertest {
 	BoardMapper mapper;
 	@Test
 	public void test() {
-		int count = mapper.count();
+		long count = mapper.count();
 		System.out.println(count);
 	}
